Updated sleep routine

+++ src/main.cpp
@@ -89,15 +89,15 @@ void frame_grabber(CCameraUnit *cam, uint64_t cadence = 10) // cadence in second
         start = get_msec() - start;
         if (start < cadence * 1000)
         {
-            int res = (start - cadence * 1000) * 1000;
-            dbprintlf("Res: %d", res);
+            uint64_t res = (start - cadence * 1000) * 1000;
+            dbprintlf("Res: %" PRIu64, res);
             usleep(res % 1000000);
             res -= res % 1000000;
             while ((res > 0) && (!done))
             {
                 usleep(1000000);
                 res -= 1000000;
-                dbprintlf("Res: %d", res);
+                dbprintlf("Res: %" PRIu64, res);
             }
         }
     }
